#20/login-frontend
Implemented a functional design for the whole base of the website. Updated the version of bootstrap from 4.3 to 5.1. Somewhat finished the design of login frontend with small neat features. But still have to swap the form from the login page to a newer one. That can talk to the backend at some point. Created a _root.css for later use of global css variables to be used on multiple styles across all elements. At last downloaded a nuget package for BrowserLink control in order to have live updates and hot reload on the website whenever something has been changed. This library is only included in development mode.