3/3-nice-to-have-effects
Worked out a few solutions that would make the UI a bit more nicé for the client user.

Fixed a bug inside the site.css with a few wrong variable calls.
Fixed an issue inside the custom-input.css where it would use the wrong variable.
Fixed an issue in _root.css where the values were wrongly assigned to the variables.

Added an Alert Component, that way alerts can we differently displayed according to usage of the message.